Moved terminal manager & shell to their own files
