+++ apps/desktop/package.json
@@ -14,6 +14,7 @@
   },
   "dependencies": {
     "@vercel/analytics": "^1.2.2",
+    "ansi-colors": "^4.1.3",
     "ansi-escapes": "^7.0.0",
     "emulators": "^0.80.22"
   }
